+++ packages/component-library-css/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @nl-rvo/component-library-css
 
+## 4.2.2
+
+### Patch Changes
+
+- 221a118: Import loader css
+
+## 4.2.1
+
+### Patch Changes
+
+- ba6a672: Fixed table header issues and design tokens updated
+
+## 4.2.0
+
+### Minor Changes
+
+- 1dee453: Added a new componont to be used, the loader.
+
 ## 4.1.0
 
 ### Minor Changes